refactor (package) : domain 패키지 내 ladder 패키지 추가
