+++ src/main/java/ladder/LadderGameApplication.java
@@ -5,13 +5,13 @@ import java.util.Map;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import ladder.domain.Height;
-import ladder.domain.Ladder;
+import ladder.domain.ladder.Height;
+import ladder.domain.ladder.Ladder;
 import ladder.domain.LadderGame;
 import ladder.domain.LadderGameResult;
-import ladder.domain.linegenerator.RandomBooleanSupplier;
-import ladder.domain.linegenerator.StickListGenerator;
-import ladder.domain.linegenerator.SticksPatternGenerator;
+import ladder.domain.ladder.linegenerator.RandomBooleanSupplier;
+import ladder.domain.ladder.linegenerator.StickListGenerator;
+import ladder.domain.ladder.linegenerator.SticksPatternGenerator;
 import ladder.domain.player.Player;
 import ladder.domain.player.Players;
 import ladder.domain.product.Product;
