+++ src/main/java/ladder/domain/ladder/linegenerator/StickListGenerator.java
@@ -1,7 +1,7 @@
-package ladder.domain.linegenerator;
+package ladder.domain.ladder.linegenerator;
 
 import java.util.List;
-import ladder.domain.Stick;
+import ladder.domain.ladder.Stick;
 
 public interface StickListGenerator {
     List<Stick> generate(int countOfPlayers);
