+++ src/test/java/ladder/domain/ladder/StickTest.java
@@ -1,7 +1,8 @@
-package ladder.domain;
+package ladder.domain.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import ladder.domain.ladder.Stick;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
